Fix the HUD saving screen to load earlier and fix a previous warning.
The saving HUD was showwing up too late so I made it appear as soon as
you press save but the previous warning
_WebThreadLockFromAnyThread(bool) came up
again. The UITextField descriptField was causing this error. When you
show HUD, you need to clear the main thread so HUD can update the UI,
the UITextField was being updated at the same time.