Lazy load our table cell thumbnail images
 We will LazyLoad the thumbnail images- meaning we will load the images
asynchronously so the table will be more responsive i.e. the table will
not delay if the images are not ready. Parse allows you to set the
PFTableViewCell's property value PFImageView imageView and all you had
to do was set the cell.imageView.file and the "loadInBackground" and
"setImage" was done for you. I used my own TableCell implementation
cell.thumbnailView which loaded the thumb images faster and I wrote out
loadInBackground and setImage so I can see clearly how the thumbnnail
image are being set.