Changed the way our deals are saved and searched
Since iOS 6 came out, the mapping data is not as accurate as Google
maps, most notably CLPlacemark locality, so now I will save each deal
in a specific postal code, instead of locality. When the user enters an
address for us to search, it will utlimately search for a specific
postal code as well.